feature: typed identifiers part 2
This change introduces the `Identifier` abstract base record class.
The `Identifier` is capable of being parsed from a string, and also
contains a defined place for storing a human-readable identifier
prefix.

An ADR (0001) was created to explain the rationale around using
typed identifiers and relates the decion back to the project
manifesto.